@@ -1,8 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
+using System.Text.Json;
 
 namespace Microsoft.EntityFrameworkCore.Internal
 {
@@ -47,19 +46,8 @@ namespace Microsoft.EntityFrameworkCore.Internal
         public int TableMaxLength { get; set; } = 128;
 
         /// <summary>
-        /// The JsonSerializerSettings for the changed column.
+        /// The json setting for the 'Changed' column
         /// </summary>
-        public JsonSerializerSettings JsonSerializerSettings { get; set; } = new JsonSerializerSettings()
-        {
-            ReferenceLoopHandling = ReferenceLoopHandling.Ignore,
-            ContractResolver = new CamelCasePropertyNamesContractResolver(),
-            Formatting = Formatting.None,
-            NullValueHandling = NullValueHandling.Ignore
-        };
-
-        /// <summary>
-        /// The json serializer to use when writing changes. Created internally.
-        /// </summary>
-        internal JsonSerializer JsonSerializer { get; set; }
+        public JsonSerializerOptions JsonSerializerOptions;
     }
 }
